+++ lib/core/versions/1.0/models/AnchoredData.ts
@@ -1,0 +1,7 @@
+/**
+ * Defines model for the data which is to be anchored to the blockchain.
+ */
+export default interface AnchoredData {
+  coreIndexFileUri: string;
+  numberOfOperations: number;
+}