+++ lib/core/versions/1.0/models/OperationReferenceModel.ts
@@ -1,0 +1,7 @@
+/**
+ * Defines the update/recover/deactivate operation reference structure found in the `operations` property in core and provisional index file.
+ */
+export default interface OperationReferenceModel {
+  didSuffix: string,
+  revealValue: string
+}