+++ lib/core/versions/1.0/models/ProvisionalIndexFileModel.ts
@@ -1,0 +1,14 @@
+import OperationReferenceModel from './OperationReferenceModel';
+
+/**
+ * Defines the external Map File structure.
+ */
+export default interface ProvisionalIndexFileModel {
+  provisionalProofFileUri?: string;
+  operations?: {
+    update: OperationReferenceModel[]
+  };
+  chunks: {
+    chunkFileUri: string
+  }[];
+}