+++ lib/core/versions/latest/util/JsonAsync.ts
@@ -7,6 +7,7 @@ export default class JsonAsync {
   /**
    * Parses the given operation into a JavaScript object asynchronously,
    * to allow the event loop a chance to handle requests.
+   * NOTE: Use only when the JSON data buffer is expected to exceed 200 KB as shown in statistics in https://github.com/ibmruntimes/yieldable-json
    */
   public static async parse (jsonData: Buffer | string): Promise<any> {